Split plugin list view model and row into separate files
