+++ WordPress/Classes/ViewRelated/Plugins/PluginListViewController.swift
@@ -1,66 +1,5 @@
-import Gridicons
 import UIKit
 import WordPressKit
-
-struct PluginListRow: ImmuTableRow {
-    static let cell = ImmuTableCell.class(WPTableViewCellSubtitle.self)
-    let name: String
-    let version: String?
-    let action: ImmuTableAction? = nil
-
-    func configureCell(_ cell: UITableViewCell) {
-        WPStyleGuide.configureTableViewSmallSubtitleCell(cell)
-        cell.textLabel?.text = name
-        cell.detailTextLabel?.text = version
-        cell.selectionStyle = .none
-        cell.imageView?.image = Gridicon.iconOfType(.plugins)
-    }
-}
-
-enum PluginListViewModel {
-    case loading
-    case ready([PluginState])
-    case error(String)
-
-    var noResultsViewModel: WPNoResultsView.Model? {
-        switch self {
-        case .loading:
-            return WPNoResultsView.Model(
-                title: NSLocalizedString("Loading Plugins...", comment: "Text displayed while loading plugins for a site")
-            )
-        case .ready:
-            return nil
-        case .error:
-            let appDelegate = WordPressAppDelegate.sharedInstance()
-            if (appDelegate?.connectionAvailable)! {
-                return WPNoResultsView.Model(
-                    title: NSLocalizedString("Oops", comment: ""),
-                    message: NSLocalizedString("There was an error loading plugins", comment: ""),
-                    buttonTitle: NSLocalizedString("Contact support", comment: "")
-                )
-            } else {
-                return WPNoResultsView.Model(
-                    title: NSLocalizedString("No connection", comment: ""),
-                    message: NSLocalizedString("An active internet connection is required to view plugins", comment: "")
-                )
-            }
-        }
-    }
-
-    func tableViewModelWithPresenter(_ presenter: ImmuTablePresenter?) -> ImmuTable {
-        switch self {
-        case .loading, .error:
-            return .Empty
-        case .ready(let pluginStates):
-            let rows = pluginStates.map({ pluginState in
-                return PluginListRow(name: pluginState.name, version: pluginState.version)
-            })
-            return ImmuTable(sections: [
-                ImmuTableSection(rows: rows)
-                ])
-        }
-    }
-}
 
 class PluginListViewController: UITableViewController, ImmuTablePresenter {
     fileprivate lazy var handler: ImmuTableViewHandler = {
